Move compile route to different folder
